refactor mapping funcs to objects
